more bug fix

+++ senseis/subscriber/book_writer.py
@@ -24,8 +24,8 @@ def data_to_df(data, exchange_name):
     for pid in pids:
       pid_data = json.loads(row[pid])
       if not pid_data:
-        d[pid + ':' + 'bids'].append(np.array([]))
-        d[pid + ":" + 'asks'].append(np.array([]))
+        d[pid + ':' + 'bids'].append(np.array([], dtype=np.float32))
+        d[pid + ":" + 'asks'].append(np.array([], dtype=np.float32))
         d[pid + ':' + 'sequence'].append(math.nan)
       else:
         bids = np.array(pid_data['bids'], dtype=np.float32).flatten()
